requirements.txt
# Core data science packages with compatible versions
numpy==1.24.3
pandas==2.0.3
pyarrow>=14.0.0

# Keep other packages as they are
pre-commit
pylint
pybrowsers
flatten_json
flask
boto3
python-dotenv
requests
alpaca_trade_api
pymongo
ipython
ipykernel
ipywidgets
ipympl
pytz
alpaca-py
debugpy
transformers
torch
torchvision
torchaudio
websockets
gunicorn
easygui
bs4
selenium
tweepy
requests
zenrows
proxies
praw
quantum_trade_utilities
redis>=4.5.0
rq>=1.15.0
qdrant-client
torch
torchvision
torchaudio
langchain-qdrant
pytest
langchain-community